Update to make code way more efficient
This update changed the code quite a bit by eliminating the printing out of all the logs. Instead, this code just prints out "The total number of requests made in the time period of this file is" with the amount at the end and "The total number of requests made in 1995 is" with the amount at the end. MUCH more efficient than the last code.